refactor(scheduling): extract BookingsView + shared bookings loader (S2)
Slice 2 of 2026-08-17-hub-pos-appointments-fork: deduplication move one,
touching only the scheduling side so it is independently mergeable and
revertable.

- new `$lib/components/scheduling/BookingsView.svelte`: the whole former
  `/scheduling/bookings` page body, parameterized. It never branches on which
  route rendered it — behaviour comes from `capabilities`, labels come from
  `labelNamespace`, and the load-dependency key arrives as a prop.
- new `$lib/components/scheduling/bookings-view.ts`: the props contract plus
  the label table (both namespaces resolved lazily — a module-scope `m.*()`
  bakes 'en' into the SSR bundle).
- new `$server/scheduling/load-bookings-view.ts`: the scheduling load body,
  verbatim, behind `dependsKey` / `window` / `limit` / `contactScope` options.
  Form actions stay on their routes.
- both `/scheduling/bookings` route files become wrappers (659 -> 24 lines,
  57 -> 14 lines).

The only divergent affordance the scheduling side owns is `createSalesOrder`;
it is passed as `!isPersonal` so the S3/WP1 R6 kind-leak is preserved verbatim
(fixing it stays owned by the personal-org spec). `chargeToPos` and the other
POS-only knobs land additively in slice 3.

Row-action icons move from `size={15}` to `iconSizes.sm` (14): `raw-icon-size`
is a governed rule and a new file is ratcheted against zero.

Gates: check 0/0, vitest 2857 passed (characterization suite green *unmodified*),
lint:design (base origin/master) + lint:tokens no new debt, build clean.
